pyproject.toml
[build-system]
requires = ["meson>=1.0.0", "meson-python>=0.15.0", "ninja", "numpy"]
build-backend = "mesonpy"

[project]
name = "SurfaceTopography"
description = "Read and analyze surface topographies"
readme = "README.md"
license = { file = "LICENSE.md" }
authors = [
    { name = "Kai Haase", email = "k.o.haase@googlemail.com" },
    { name = "Johannes Hörmann", email = "johannes.hoermann@imtek.uni-freiburg.de" },
    { name = "Till Junge", email = "till.junge@altermail.ch" },
    { name = "Lars Pastewka", email = "lars.pastewka@imtek.uni-freiburg.de" },
    { name = "Michael Röttger", email = "info@michael-roettger.de" },
    { name = "Antoine Sanner", email = "antoine.sanner@imtek.uni-freiburg.de" }
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Programming Language :: Python"
]
requires-python = ">=3.8.0"
dynamic = [ "version" ]
dependencies = [
    "numpy",
    "scipy>=1.9.0",
    "defusedxml",
    "h5py",
    "igor2",
    "matplotlib>=1.0.0",
    "muSpectre>=0.25.1",
    "NuMPI>=0.3.1",
    "numpyencoder",
    "openpyxl",
    "pandas",
    "Pillow",
    "python-dateutil",
    "pyyaml",
    "requests",
    "tiffile",
    "xmltodict",
    "olefile"
]

[project.optional-dependencies]
test = [
    "flake8<5",
    "pytest",
    "pytest-cov",
    "pytest-flake8",
    "runtests"
]

[project.urls]
documentation = "https://contactengineering.github.io/SurfaceTopography/"
repository = "https://github.com/ContactEngineering/SurfaceTopography"
changelog = "https://github.com/ContactEngineering/SurfaceTopography/blob/master/SurfaceTopography/ChangeLog.md"
